requirements/base.txt
django==1.9.1

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.6

# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.6.0

# Models
django-model-utils==2.3.1

# Images
# Pillow==3.0.0

# For user registration, either via email or social
django-allauth==0.24.1

# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
# django-redis==4.3.0
# redis>=2.10.0

django-suit==0.2.15
# django-material
python-dateutil

# temp get master as I need xlsx support
# https://github.com/django-import-export/django-import-export/archive/master.zip
django-import-export
openpyxl
xlrd
unipath
beautifulsoup4
